いいね機能実装完了

+++ app/controllers/favorites_controller.rb
@@ -1,15 +1,21 @@
 class FavoritesController < ApplicationController
   def create
     tweet = Tweet.find(params[:tweet_id])
-    favorite = current_user.favorites.new(tweet_id: tweet.id)
-    favorite.save
-    redirect_to tweet_path(tweet)
+    if user_signed_in?
+      favorite = current_user.favorites.new(tweet_id: tweet.id)
+      favorite.save
+    else
+      redirect_to tweet_path(tweet)
+    end
   end
   
   def destroy
     tweet = Tweet.find(params[:tweet_id])
-    favorite = current_user.favorites.find_by(tweet_id: tweet.id)
-    favorite.destroy
-    redirect_to tweet_path(tweet)
+    if user_signed_in?
+      favorite = current_user.favorites.find_by(tweet_id: tweet.id)
+      favorite.destroy
+    else
+      redirect_to tweet_path(tweet)
+    end
   end
 end
